Split generic bound validation
