+++ src/typechecker/mod.rs
@@ -17,6 +17,7 @@ mod declaration_collection;
 mod declaration_collection_ast;
 mod environment;
 mod expressions;
+mod generic_bound_validation;
 mod generic_type_validation;
 mod monomorphize;
 mod patterns;
